+++ src/components/calendar/DateBoxList.tsx
@@ -1,7 +1,3 @@
-import React, { useContext, useMemo } from 'react'
-import { useStore } from 'zustand'
-
-import { CalendarContext } from '@/store/calendarStore'
 import { validateDateBoxColor } from '@/utils/validateTheme'
 
 import { DateBox } from './DateBox'
@@ -15,26 +11,20 @@ const DateBoxColors = {
   default: '',
 }
 
-export const DateBoxList = () => {
-  const store = useContext(CalendarContext)
-  if (!store) throw new Error('Missing CalendarContext.Provider in the tree')
-  const date = useStore(store, (s) => s.date)
-  const schedules = useStore(store, (s) => s.schedules)
+type DateBoxListProps = {
+  date: Date
+  schedules: string[]
+  datesInMonth: number
+}
 
-  const parseSchedules = useMemo(
-    () => schedules?.map((schedule) => schedule.split(' ')[0]),
-    [schedules],
-  )
-
+export const DateBoxList = ({ date, schedules, datesInMonth }: DateBoxListProps) => {
   return (
     <>
-      {Array.from({ length: new Date(date.getFullYear(), date.getMonth() + 1, 0).getDate() }).map(
-        (_, idx) => {
-          const dayIndex = idx + 1
-          const color = validateDateBoxColor(date, parseSchedules || [], dayIndex)
-          return <DateBox key={dayIndex} num={dayIndex} color={DateBoxColors[color]} />
-        },
-      )}
+      {Array.from({ length: datesInMonth }).map((_, idx) => {
+        const dayIndex = idx + 1
+        const color = validateDateBoxColor(date, schedules, dayIndex)
+        return <DateBox key={dayIndex} num={dayIndex} color={DateBoxColors[color]} />
+      })}
     </>
   )
 }
